requirements.txt
asgiref==3.8.1
astroid==3.3.8
dill==0.3.9
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
pylint==3.3.3
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.8.2
